requirements.txt
aiofiles
asyncpg
python_dotenv
fastapi
pydantic
psycopg2-binary
uvicorn